+++ src/main/java/lotto/model/domain/LottoMachine.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
