+++ src/main/java/lotto/model/domain/PurchaseAmount.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model.domain;
 
 public class PurchaseAmount {
     private int purchaseAmount;
