+++ src/test/java/lotto/model/domain/BonusNumberTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
