+++ src/test/java/lotto/model/domain/LottosTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model.domain;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
